Updated architecture diagram to include reverse vpn tunnel and sni infrastructure.
Added vpn server/envoy to the control plane in the seed and removed the vpn load balancer in the shoot. Furthermore, removed the vpn side-car in the monitoring of the seed. Renamed the shoot api server load balancer to istio ingress load balancer as it is now used for the sni infrastructure.