Improve alert documentation for shoot owners (#532)
The list of predefined alerts for shoot owners might change frequently
and make this documentation outdated. For instance, it is not true
anymore that the number of predefined alerts is 16.

This commit improves this documentation by writing it in a way that does
not break so easy. First, the number of predefined alerts is avoided.
Second, it explains how to find the list of predefined alerts from the
shoot Prometheus UI.